updated string file
